+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Added command `odoo-helper install dev-tools` that is just an alias to install
   *bin-tools*, *py-tools* and *js-tools* with single command.
+- Param `--db` to `odoo-helper addons find-installed` to search for addons only
+  in specified databases.
 
 ### Changed